engine: extract esm/render_context.js
